+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - [x] Obstacles no longer require HMs to be known by a party member, just the HM in the player's bag.
 - [x] TMs are now infinitely reuseable. You can only get one copy of each TM.
 - [ ] Natures now have stat colours in the Summary Screen.
-- [ ] EVs and IVs are now shown in the Summary Screen.
+- [x] EVs and IVs are now shown in the Summary Screen.
 - [x] Reversed Running Shoes behaviour. Press B to walk.
 - [x] Text Speed now defaults to Fast.
 - [x] L & R now turbos A & B buttons under L=A option mode.